tools/requirements.txt
click==8.1.7
greenlet==3.0.0
numpy==1.26.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
shapely==2.0.2
SQLAlchemy==2.0.22
typing_extensions==4.8.0
